Refactor tests./ Split between local ones and ones needing GCP
Note: some tests needing GCP seem to be broken already (e.g. test:gcp-auth.js)
