+++ package.json
@@ -9,14 +9,14 @@
   },
   "scripts": {
     "deploy": "gcloud run deploy cloud-run-mcp --source . --no-invoker-iam-check",
-    "test:deploy": "node test/test-deploy.js",
-    "test:create-project": "node test/test-create-project.js",
-    "test:service-logs": "node test/test-service-logs.js",
-    "test:gcp-auth": "node test/test-gcp-auth.js",
-    "test:prompts": "node --test test/test-prompts.js",
-    "test:mcp": "node --test test/test-mcp-server.js",
-    "test:tools": "node --test test/test-tools.js",
-    "test": "node test/run-all-tests.js",
+    "test:deploy": "node test/need-gcp/cloud-run-deploy.test.js",
+    "test:create-project": "node test/need-gcp/gcp-projects.test.js",
+    "test:service-logs": "node test/need-gcp/cloud-run-services.test.js",
+    "test:gcp-auth": "node test/need-gcp/gcp-auth-check.test.js",
+    "test:prompts": "node --test test/local/prompts.test.js",
+    "test:mcp": "node --test test/local/mcp-server.test.js",
+    "test:tools": "node --test test/local/tools.test.js",
+    "test": "node --test test/local/*.test.js",
     "mcp-inspector": "npx @modelcontextprotocol/inspector node mcp-server.js",
     "start": "node mcp-server.js",
     "lint:check": "prettier --check .",
