+++ test/local/mcp-server.test.js
