+++ test/local/tools.test.js
@@ -8,7 +8,7 @@ describe('registerTools', () => {
       registerTool: mock.fn(),
     };
 
-    const { registerTools } = await esmock('../tools.js', {});
+    const { registerTools } = await esmock('../../tools.js', {});
 
     registerTools(server);
 
@@ -38,8 +38,8 @@ describe('registerTools', () => {
         registerTool: mock.fn(),
       };
 
-      const { registerTools } = await esmock('../tools.js', {
-        '../lib/gcp-projects.js': {
+      const { registerTools } = await esmock('../../tools.js', {
+        '../../lib/gcp-projects.js': {
           listProjects: () =>
             Promise.resolve([{ id: 'project1' }, { id: 'project2' }]),
         },
@@ -69,8 +69,8 @@ describe('registerTools', () => {
         registerTool: mock.fn(),
       };
 
-      const { registerTools } = await esmock('../tools.js', {
-        '../lib/gcp-projects.js': {
+      const { registerTools } = await esmock('../../tools.js', {
+        '../../lib/gcp-projects.js': {
           createProjectAndAttachBilling: (projectId) =>
             Promise.resolve({
               projectId: projectId,
@@ -101,8 +101,8 @@ describe('registerTools', () => {
         registerTool: mock.fn(),
       };
 
-      const { registerTools } = await esmock('../tools.js', {
-        '../lib/gcp-projects.js': {
+      const { registerTools } = await esmock('../../tools.js', {
+        '../../lib/gcp-projects.js': {
           createProjectAndAttachBilling: () =>
             Promise.resolve({
               projectId: 'generated-project',
@@ -135,8 +135,8 @@ describe('registerTools', () => {
         registerTool: mock.fn(),
       };
 
-      const { registerTools } = await esmock('../tools.js', {
-        '../lib/cloud-run-services.js': {
+      const { registerTools } = await esmock('../../tools.js', {
+        '../../lib/cloud-run-services.js': {
           listServices: () =>
             Promise.resolve([
               { name: 'service1', uri: 'uri1' },
@@ -172,8 +172,8 @@ describe('registerTools', () => {
         registerTool: mock.fn(),
       };
 
-      const { registerTools } = await esmock('../tools.js', {
-        '../lib/cloud-run-services.js': {
+      const { registerTools } = await esmock('../../tools.js', {
+        '../../lib/cloud-run-services.js': {
           getService: () =>
             Promise.resolve({
               name: 'my-service',
@@ -223,8 +223,8 @@ describe('registerTools', () => {
         return Promise.resolve({ logs: 'log3\nlog4', requestOptions: null });
       };
 
-      const { registerTools } = await esmock('../tools.js', {
-        '../lib/cloud-run-services.js': {
+      const { registerTools } = await esmock('../../tools.js', {
+        '../../lib/cloud-run-services.js': {
           getServiceLogs: getServiceLogs,
         },
       });
@@ -257,8 +257,8 @@ describe('registerTools', () => {
         registerTool: mock.fn(),
       };
 
-      const { registerTools } = await esmock('../tools.js', {
-        '../lib/cloud-run-deploy.js': {
+      const { registerTools } = await esmock('../../tools.js', {
+        '../../lib/cloud-run-deploy.js': {
           deploy: () => Promise.resolve({ uri: 'my-uri' }),
         },
       });
@@ -292,8 +292,8 @@ describe('registerTools', () => {
         registerTool: mock.fn(),
       };
 
-      const { registerTools } = await esmock('../tools.js', {
-        '../lib/cloud-run-deploy.js': {
+      const { registerTools } = await esmock('../../tools.js', {
+        '../../lib/cloud-run-deploy.js': {
           deploy: () => Promise.resolve({ uri: 'my-uri' }),
         },
       });
@@ -327,8 +327,8 @@ describe('registerTools', () => {
         registerTool: mock.fn(),
       };
 
-      const { registerTools } = await esmock('../tools.js', {
-        '../lib/cloud-run-deploy.js': {
+      const { registerTools } = await esmock('../../tools.js', {
+        '../../lib/cloud-run-deploy.js': {
           deploy: () => Promise.resolve({ uri: 'my-uri' }),
         },
       });
@@ -362,8 +362,8 @@ describe('registerTools', () => {
         registerTool: mock.fn(),
       };
 
-      const { registerTools } = await esmock('../tools.js', {
-        '../lib/cloud-run-deploy.js': {
+      const { registerTools } = await esmock('../../tools.js', {
+        '../../lib/cloud-run-deploy.js': {
           deployImage: () => Promise.resolve({ uri: 'my-uri' }),
         },
       });
